requirements.txt
pandas 
re
pymongo
python-dotenv
pymupdf
tqdm
langchain
langchain-community
fastapi
uvicorn
-e .